requirements.txt
# Минимальные зависимости для Vercel деплоя
python-telegram-bot==20.7
requests==2.31.0
